+++ packages/components/src/private-apis.ts
@@ -16,9 +16,12 @@ import {
 	ValidatedCheckboxControl,
 	ValidatedInputControl,
 	ValidatedNumberControl,
+	ValidatedSelectControl,
+	ValidatedRadioControl,
 	ValidatedTextControl,
 	ValidatedTextareaControl,
 	ValidatedToggleControl,
+	ValidatedToggleGroupControl,
 } from './validated-form-controls';
 import { ValidatedFormTokenField } from './validated-form-controls/components/form-token-field';
 import { Picker } from './color-picker/picker';
@@ -41,8 +44,11 @@ lock( privateApis, {
 	ValidatedInputControl,
 	ValidatedCheckboxControl,
 	ValidatedNumberControl,
+	ValidatedSelectControl,
+	ValidatedRadioControl,
 	ValidatedTextControl,
 	ValidatedTextareaControl,
 	ValidatedToggleControl,
-	ValidatedFormTokenField,
+	ValidatedToggleGroupControl,
+  ValidatedFormTokenField,
 } );