+++ packages/fields/src/actions/rename-post.tsx
@@ -21,12 +21,7 @@ import { store as noticesStore } from '@wordpress/notices';
  */
 
 import { unlock } from '../lock-unlock';
-import {
-	getItemTitle,
-	isTemplateRemovable,
-	isTemplate,
-	isTemplatePart,
-} from './utils';
+import { getItemTitle, isTemplatePart } from './utils';
 import type { CoreDataError, PostWithPermissions } from '../types';
 
 // Patterns.
@@ -43,21 +38,11 @@ const renamePost: Action< PostWithPermissions > = {
 		// Templates, template parts and patterns have special checks for renaming.
 		if (
 			! [
-				'wp_template',
 				'wp_template_part',
 				...Object.values( PATTERN_TYPES ),
 			].includes( post.type )
 		) {
 			return post.permissions?.update;
-		}
-
-		// In the case of templates, we can only rename custom templates.
-		if ( isTemplate( post ) ) {
-			return (
-				isTemplateRemovable( post ) &&
-				post.is_custom &&
-				post.permissions?.update
-			);
 		}
 
 		if ( isTemplatePart( post ) ) {